* move PrintReport() and dependent parts into own file
