+++ src/reporting.h
@@ -1,0 +1,12 @@
+#ifndef REPORTING_H_JHA2MUHM
+#define REPORTING_H_JHA2MUHM
+
+#include "SourceAnnotation.h"
+
+
+void PrintReport(const snowcrash::Report& report,
+                 const std::string& source,
+                 const bool isUseLineNumbers);
+
+
+#endif /* end of include guard: REPORTING_H_JHA2MUHM */
